moved SVG test images to own class
